network/zabbix_agentd: Updated for version 1.8.16.
This commit was backported from the 14.0 repo at the request
of the maintainer.  The rationale is that lots of people are
still using zabbix*-1.8.x, and since the 14.0 repo will soon
be updated for 2.x, this would give those people the latest
stable version with a minimum of hassle for them.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>